basic/save_tf_model/requirements.txt
tensorflow==2.12.0
numpy>=1.19.2
pyyaml==6.0.2